Add files via upload
Here the objective was to merge the two tables named "sales" and "product" using the merge function in Power BI. In order to execute this, you have to go to transform data in the query editor and then import the .xlsx files into it. After that, go to transform data; under the transform tab, go to merge under the merge option. hit create a new merge option" and then select the merge then merge it